Moved and renamed code related to the Inputs Dialog and related UX
Primarily the following:
* `crate::dialogs::input::InputDialog` ==> `crate::dialogs::input::primary::InputDialog`
* `crate::dialogs::input_multi::InputMultiDialog` ==> `crate::dialogs::input::xy::InputXyDialog`
* `crate::appcommand::AppCommand::InputMultiDialog` ==> `crate::appcommand::AppCommand::InputXyDialog`

Also moved shared code to `crate::dialogs::input`.

Calling the "XY" dialog "Multi" is bad when we are about to implement the "Multiple..." context menu item
