+++ ui/main.slint
@@ -9,7 +9,7 @@ import { DeviceAndImageEntry } from "devimageslist.slint";
 import { AppWindow } from "appwindow.slint";
 import { ConfigureDialog } from "configure.slint";
 import { InputDialog, InputDialogEntry } from "input.slint";
-import { InputMultiDialog } from "input_multi.slint";
+import { InputXyDialog } from "input_xy.slint";
 import { SeqPollDialog } from "seqpoll.slint";
 
-export { AboutDialog, AppWindow, ConfigureDialog, ConnectToSocketDialog, MessageBoxDialog, NameCollectionDialog, PathsDialog, DevicesAndImagesDialog, DeviceAndImageEntry, InputDialog, InputDialogEntry, InputMultiDialog, SeqPollDialog, Icons }
+export { AboutDialog, AppWindow, ConfigureDialog, ConnectToSocketDialog, MessageBoxDialog, NameCollectionDialog, PathsDialog, DevicesAndImagesDialog, DeviceAndImageEntry, InputDialog, InputDialogEntry, InputXyDialog, SeqPollDialog, Icons }
